OF-1246 Update Monitoring plugin to support MAM-MUC
Note that this also fixes numerous bugs with general MAM itself. A full
comparitive test run was not performed, so it is unclear how many issues
were present in previous builds.

The "with" field is mandatory within MAM, but is not within MAM-MUC; it is
therefore ignored. The specification remains unclear on whether this is
legitimate.