38: Try to deduce Unique/Stable ID when using MAM2
When being queried through MAM2, the RSM-provided identifiers are expected to be Unique/Stable IDs (XEP-0359). With this change, these values are translated into the database identifiers that Openfire uses internally.

This commit causes costly database queries to be performed. It's likely that in the future, the XEP-0359 identifiers get a dedicated column in a database table, which should improve performance.