38: Various fixes
A change has been included that forces a ResultSetManagement object to be used, preventing a NullPointerException in certain cases. Forcing a RSM instruction when none is present is enabled by default, but can be turned off by setting the value of property 'archive.FORCE_RSM' to false.

An earlier commit introduced the usage of unique and stable IDs for RSM pagination. Clients that have existing data will not immediately use this, but will use the old database identifier that they previously got when retrieving history. This commit falls back to this mechanism, when it's determined that the old identifiers are used by the client.

Resolved a bug where the state of 'is paging backwards' was not detected properly. Moved the functionality to the RSM implementation.