fixes #58: Poor performance when adding messages to a large database.
This commit replaces the mechanism to determine the next (unique) messageID value for rows in the ofMessageArchive table.

Instead of using count(*) to determine the next value, a sequence is used (as provided by Openfire's SequenceManager).

A database update script has been included that automatically initializes the sequence to the correct value, for databases that already have content.