fixes #62: Set 'complete' only on last page retrieval
This commit splits the code that performs the query from the arguments used from paging through the result set - at least for the Lucene based MAM MUC implementation. The new code is utilized to 'get the next page' in order to determine if there is such a next page, for the same query. If there isn't then the 'complete' flag can be set on the result.