fixes #80: Embedded database only: Prevent all messages to be returned when only a count is requested.
Not querying the database for messages if you're requesting zero messages to be returned makes sense from a performance perspective. Additionally, the existing code trips what appears to be a bug in HSQLDB which appears to return all rows (instead of no rows) if you end a query with "LIMIT 0".