fixes #70: Purge null RRDs from database when starting.
Apparently, something occasionally corrupts the database table that stores the RRDS holding statistics. When one row is corrupt, it affects the functionality of all statistics.

To work around this issue, this commit removes all affected rows upon start of the plugin. This will cause the code to recreate the row with a valid structure, and re-enables usage of all other statistics.