fixes #84: Return error on unrecognized MAM query fields.
Some clients use a data form field that is unrecognized. Up until now, the monitoring plugin has ignored those fields. This can lead to unexpected results, as the client (and end user) assume that a certain limitation is applied to the search, while in fact, it is not.

The plugin should return an error when unsupported fields are used, instead of silently ignoring them.

This behavior can be configured by using the 'monitoring.search.allow-unrecognized-fields' property.