fixes #98: RSM results must have 'first' and 'last' for non-empty results.
This fixes a bug where the 'first' and 'last' element values of the RSM element were not set when the corresponding messages do not contain a unique and stable message ID (which were introduced as part of MAM:2 support).

The solution implemented here falls back to using the old-style message-id, which is equally usable to retrieve messages.