fixes #101: Do not assume XEP-0359 values to be UUIDs
Although the XEP specifies that XEP-0359 identifiers SHOULD be UUIDs, they need not be. At least one third party implementation (converse.js) is known to not use UUIDs.

This change replaces the UUID format with a regular string. This will make detection of 'XEP-0359 usage' less precise. In the old code, the existence of a UUID value in the stanza text was used. To prevent any stanza to match common identifiers, an additional check is added that requires the stanza to contain the XEP-0359 namespace. Far from ideal, but until we properly spread things over queryable database columns...