fixes #113: Database migration to separate off resource-part of JIDs
Database update script 1 of the monitoring plugin introduces new columns to the ofMessageArchivePlugin. These new olumn are intended to hold the optional resource-part of a JID, while the original column, that previously held he full JID, now intends to hold the bare JID.

The update script does not perform migration of existing data. This migration is performed by this method. The functionality in this method is idempotent. However, as execution can be costly, especially on larger data sets, even when there's no work to perform, a property is set after the first successful execution. This property prevents subsequent executions.

Additionally, this method will skip execution when the code is running in a cluster and the local node is not the senior node. This intends to prevent multiple nodes from attempting to make the same database changes simultaneously.