fixes #146: Add option to choose between database used for MUC archive
The archived MUC messages are persisted to the database more than once: Openfire stores them in ofMucConversationLog, the Monitoring plugin in ofMessageArchive.

This commit introduces a configuration option to switch between the two: conversation.database.use-openfire-tables

The default for this new option is to use the database table provided by the Monitoring plugin, which is a change from the behavior prior to this commit.