fixes #147: Allow how MUC PMs are stored to be configurable
Historically, Private Messages sent through MUC rooms are stored as if they were one-on-one messages (in the personal archives of users). With #133 these PMs are now also stored in the MUC archive.

Existing functionality might depend on these messages to be stored in the personal archives, or might break when these messages start showing up in the MUC archives.

The changes in #133 retain the old functionality, but also adds the messages to the MUC archives (there's duplication now).

Configuration options should be made available to control this behavior.

Two new properties to be introduced, both take boolean values, both default to true:

- conversation.roomArchiving.PMinPersonalArchive
- conversation.roomArchiving.PMinRoomArchive