fixes #159: Remove Picocontainer
The monitoring plugin uses the third party Picocontainer library, but uses only a fraction of its functionality. In heap dumps created to diagnose memory leaks, classes from this library are visible. It is likely that Picocontainer isn't actually contributing to the problem, but to reduce complexity, we might as well replace Picocontainer with a simple start/stop lifecycle implementation. This would bring the code in line with most other plugins and Openfire implementation (none of which use Picocontainer).