fixes #160: Update JRobin
The monitoring plugin uses the third party JRobin library for its RRD implementation used to collect statistics.

In heap dumps created to diagnose memory leaks, classes from this library are visible. It is likely that JRobin isn't actually contributing to the problem, but we might as well update JRobin, as later releases do seem to include fixes that relate to instance reference management.

JRobin's later releases aren't available in the central Maven repository, but are provided publicly in the OpenNMS repository.