fixes #155: Allow ThreadPoolExecutor's threads to be garbage collected quickly.
The root cause of #155 can be found in the PluginClassLoader that loaded the plugin not being garbage collected 'soon enough'. It has been observed that the PluginClassLoader remained in memory for approximately one minute. When within that time a replacement plugin is loaded, ClassCastExceptions occur, as both ClassLoaders would have loaded the same class.

The 60 second duration is equal to the amount of time that various ExecutorServices allow their worker threads to be idle, before being cleaned up. This is the default behavior of ExecutorService instances returned by `java.util.concurrent.Executors#newCachedThreadPool(java.util.concurrent.ThreadFactory)`. Experimentation found that having a lower maximum idle value would affect the speed with which the PluginClassLoader becomes eligible for garbage collection.

This commit adds a delegate for an ExecutorService that, as part of its 'shutdown' implementation, will set the maximum idle value to a very small value. This coerces the threads from being cleaned up almost instantly. This in turn prevents the classloading problems described above.