fixes #175: Fix MySQL installation script
The MySQL installation script was modified in release 2.2.0 to add a new column. This column should have been defined as NULL, but was defined as NOT NULL. The database update script (7) does is correctly (the problem only affects fresh installations that use the initial installation script).

This commit adds a new update script (number 8) that corrects the column definition for MySQL, and corrects the original installation script. It is otherwise a no-op for other databases.