fixes #125: Return results for keyword and date range combination query
When using a keyword for searching a message archive, the implementation uses the Lucene-based index, rather than the database index, to find the appropriate messages. The date-based search functionality of Lucene is broken, which is fixed by this commit. Instead of attepting to compare a byte-representation of the date of a message, a numeric representation is used instead.