fixes #190: Allow forced Lucene re-indexation
This commit adds a version number for the structure of the Lucene index that the code is expected to work with. The same version is stored in a file with the actual index whenever it is created.

When the plugin gets loaded or reloaded, the version on file will be compared by the version in code, forcing a recreation if the two do not match.