Split translations
For this plugin, the 'name' and the 'file name' differ. This inconsistency exposes an oddity with how i18n works in Openfire.

'Regular' translations (eg: admin page stuff) is retrieved based on the 'file name' of the plugin.

Translations for System Properties are based on the plugin name provided in the property. For properties to automatically unload after a plugin removal, this needs to be the 'name' of the plugin.

In this commit, I'm splitting the i18n files in two variants: the old one ('monitoring_i18n') remains the one used for regular translations. The new one ('monitoringservice_i18n') is only used for the System Properties defined in the plugin.