fixes #238: Release references provided by unloaded plugins
When another plugin provides a Statistic instance, that instance should be released when the provider unloads. Failing to do so will prevent that plugin's classloader from being destroyed. This can cause all kinds of unexpected side-effects.

This commit removes references to _all_ Statistic instances whenever _any_ plugin unloads. That's a buckshot, but the effects should be limited to missing stats around the time that a plugin gets unloaded/reloaded - which is expected to be very rare.