fixes #357: No longer show NPE on admin console after user session expired
The NPE shown was an artifact of the page being processed without authentication in place. This should be prevented by the AuthCheckFilter.

Sadly, a naming collision prevents the AuthCheckFilter to operate on the Monitoring Service admin console pages.

This commit prevents the naming collision, by renaming the endpoint on which the public API for logs is exposed (from 'monitoring' to 'logs').