Database should track room ID
The room ID differs from a room JID in the following way: when a room gets destroyed, and is recreated, it can be recreated using the same JID. The numeric identifier however, will be unique.

The room ID can therefor be used to differentiate between data recorded for various 'reincarnations' of a room with the same name.

The database update scripts will populate the room ID for existing rooms with the latest ID as recorded in the Openfire table `ofMucRoom`. This is not necessarily correct (as data recorded by the Monitoring plugin may relate to earlier versions of a room). This behavior does, however, reflect the current behavior of the Monitoring plugin, that currently always associates data with the 'latest' reincarnation of the room.

Starting with this commit, new incarnations of the room will receive distinct IDs.

This is a prerequisite for fixing issues #369 and #370.