Remove redundant StanzaID code
In relation to issues #118 and #101 a class from Openfire was duplicated in the Monitoring plugin (in commit 4065f8973946e75b2b4355ca91a99181f0acef7d ).

Now that this plugin requires a version of Openfire that has the original code, the duplicate is no longer needed. This commit replaces usage of the duplicated code with that of the original.