Prevent database storage issue when using statistics with a key that's longer than 20 characters
RRD's implementation in Java requires a data source name that's no longer than 20 characters. The key under which Openfire's Statistic definition instances are stored in the StatisticsManager are used as a RRD data source name. Thus, if that key is longer than 20 characters, an error is logged (and data points are not stored in the database).

To prevent this issue, this commit introduces a new method that ensures that the value used is not longer than 20 characters. It will use the original value if it matches that criterium, or use a 20-character hash (which is designed to be uniformly distributed) if it's longer.

As a note of caution: the hashing function was provided by AI. It _seems_ sensible to me, but it has not exhaustively tested (other than through the basic unit test that is added as part of this commit).

fixes #436