Remove 'private messages' from MUC MAM query results
XEP-0313 clearly specifies that 'private messages' MUST NOT be part of a MUC message archive (implying that it instead is part of the personal archive).

In this change, the code is refactored to no longer include 'private messages' in MAM queries that are executed against MUC archives. As a consequence, the commonly used `messageOwner` argument (which was introduced to be able to query for PMs in MUC archives) is no longer needed, and is removed.

This commit is one of several that, together, aim to fix #422